Extract commands from CLI
