+++ lib/rubocop.rb
@@ -604,6 +604,14 @@ require_relative 'rubocop/processed_source'
 require_relative 'rubocop/result_cache'
 require_relative 'rubocop/runner'
 require_relative 'rubocop/cli'
+require_relative 'rubocop/cli/command'
+require_relative 'rubocop/cli/environment'
+require_relative 'rubocop/cli/command/base'
+require_relative 'rubocop/cli/command/auto_genenerate_config'
+require_relative 'rubocop/cli/command/execute_runner'
+require_relative 'rubocop/cli/command/init_dotfile'
+require_relative 'rubocop/cli/command/show_cops'
+require_relative 'rubocop/cli/command/version'
 require_relative 'rubocop/options'
 require_relative 'rubocop/remote_config'
 require_relative 'rubocop/yaml_duplication_checker'
